requirements.txt
ultralytics
streamlit
pillow
pandas
numpy
opencv-python-headless
huggingface_hub
